temp commit

+++ src/main/java/bookshop/dao/book/CommentDao.java
@@ -7,10 +7,10 @@ import java.util.List;
 
 public interface CommentDao {
 	List<Comment> getAll(Integer bookId);
-	Comment getCommentById(Long id);
+	Comment getCommentById(Integer  id);
 	void persistComment(Integer bookId, Comment comment);
-	void deleteCommentById(Long id, Integer bookId);
-	void updateComment(Long id);
-	void likeComment(Long id, User user);
-	void dislikeComment(Long id, User user);
+	void deleteCommentById(Integer  id, Integer bookId);
+	void updateComment(Integer  id);
+	void likeComment(Integer  id, User user);
+	void dislikeComment(Integer  id, User user);
 }
