+++ src/main/java/bookshop/service/book/CommentService.java
@@ -7,10 +7,10 @@ import bookshop.model.user.User;
 
 public interface CommentService {
 	List<Comment> getAll(Integer bookId);
-	Comment getCommentById(Long id);
+	Comment getCommentById(Integer  id);
 	void persistComment(Integer bookId, Comment comment);
-	void updateComment(Long id);
-	void likeComment(Long id, User user);
-	void dislikeComment(Long id, User user);
-	void deleteCommentById(Long id, Integer bookId);
+	void updateComment(Integer  id);
+	void likeComment(Integer  id, User user);
+	void dislikeComment(Integer  id, User user);
+	void deleteCommentById(Integer  id, Integer bookId);
 }
